% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSDp.r
\name{RMSDp}
\alias{RMSDp}
\title{Modified Stahel-Donoho Estimators (paralleled version)}
\usage{
RMSDp(inp, nb = 0, sd = 0, pt = 0.999, dv = 10000)
}
\arguments{
\item{inp}{imput data (a numeric matrix)}

\item{nb}{number of basis}

\item{sd}{seed (for reproducibility)}

\item{pt}{threshold for outlier detection (probability)}

\item{dv}{maximum number of elements processed together on the same core}
}
\value{
a list of the following information
\itemize{
  \item u  final mean vector
  \item V  final covariance matrix
  \item wt final weights
  \item mah Mahalanobis distances of each observation
  \item cf threshold to detect outliers (percentile point)
  \item ot 1:normal observation, 2:outlier
}
}
\description{
This function is for multivariate outlier detection.
    Ver.3.1 2013/06/15  Related paper: DOI: 10.1109/CLOUDCOM-ASIA.2013.86
    Ver.3.2 2021/11/15  Outlier detection step added
    Ver.3.3 2022/08/12  Bug fixed about Rondom seed setting
}
