% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wine}
\alias{wine}
\title{Wine dataset in UCI Machine Learning Repository}
\format{
## `wine`
A data frame with 178 rows and 13 columns:
\describe{
  \item{Alcohol}{Alcohol}
  \item{Malic acid}{Malic acid}
  \item{Ash}{Ash}
  \item{Alcalinity of ash}{Alcalinity of ash}
  \item{Magnesium}{Magnesium}
  \item{Total phenols}{Total phenols}
  \item{Flavonoids}{Flavonoids}
  \item{Nonflavanoid phenols}{Nonflavanoid phenols}
  \item{Proanthocyanins}{Proanthocyanins}
  \item{Color intensity}{Color intensity}
  \item{Hue}{Hue}
  \item{OD280/OD315 of diluted wines}{OD280/OD315 of diluted wines}
  \item{Proline}{Proline}
}
}
\source{
<https://archive.ics.uci.edu/dataset/109/wine>
}
\usage{
wine
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
